make the attribute cap a double

+++ src/main/java/dev/amraleth/afo/item/FishingAttribute.java
@@ -61,7 +61,7 @@ public enum FishingAttribute {
     /**
      * The cap of this stat
      */
-    private final int max;
+    private final double max;
 
     /**
      * Gets all attributes from an {@link ItemStack}
